Split irdl.py
