+++ pyproject.toml
@@ -85,7 +85,10 @@ target-version = "py310"
 "xdsl.parser.attribute_parser".msg = "Use xdsl.parser instead."
 "xdsl.parser.affine_parser".msg = "Use xdsl.parser instead."
 "xdsl.ir.core".msg = "Use xdsl.ir instead."
-"xdsl.irdl.irdl".msg = "Use xdsl.irdl instead"
+"xdsl.irdl.common".msg = "Use xdsl.irdl instead"
+"xdsl.irdl.constraints".msg = "Use xdsl.irdl instead"
+"xdsl.irdl.attributes".msg = "Use xdsl.irdl instead"
+"xdsl.irdl.operations".msg = "Use xdsl.irdl instead"
 "xdsl.ir.affine.affine_expr".msg = "Use xdsl.ir.affine instead"
 "xdsl.ir.affine.affine_map".msg = "Use xdsl.ir.affine instead"
 "xdsl.ir.affine.affine_set".msg = "Use xdsl.ir.affine instead"
