+++ xdsl/irdl/__init__.py
@@ -1,3 +1,6 @@
 # TID 251 enforces to not import from nested irdl
 # We need to skip it here to allow importing from here instead.
-from .irdl import *  # noqa: TID251
+from .attributes import *  # noqa: TID251
+from .common import *  # noqa: TID251
+from .constraints import *  # noqa: TID251
+from .operations import *  # noqa: TID251
